// frontend/src/api/authentication/models/oTPRequestRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * ElectroShop API
 * E-commerce platform API with JWT authentication
 * OpenAPI spec version: 1.0.0
 */
import type { PurposeEnum } from "./purposeEnum";

/**
 * Serializer for requesting OTP with single identifier field.
 */
export interface OTPRequestRequest {
  /**
   * Email address or phone number
   * @minLength 1
   */
  identifier: string;
  /** Purpose of OTP: register or login

* `register` - register
* `login` - login */
  purpose: PurposeEnum;
}