// frontend/src/api/authentication/models/purposeEnum.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * ElectroShop API
 * E-commerce platform API with JWT authentication
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `register` - register
 * `login` - login
 */
export type PurposeEnum = (typeof PurposeEnum)[keyof typeof PurposeEnum];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PurposeEnum = {
  register: "register",
  login: "login",
} as const;